feat(client): added product client and documentation

+++ src/main/java/com/arca/order/modules/orderProduct/domain/dto/OrderProductDisplayDto.java
@@ -9,7 +9,7 @@ public record OrderProductDisplayDto(
         UUID productId,
         String name,
         Double price,
-        Double quantity
+        Integer quantity
 )
 {
     public OrderProductDisplayDto(OrderProduct orderProduct)
